+++ Blogs/Blogs/Screens/Cells/Standart Blog Cell/StandartBlogCellViewModel.swift
@@ -6,10 +6,10 @@
 //
 
 import Foundation
-import UIKit
 
 final class StandartBlogCellViewModel: StandartBlogCellIdentifiable {
     
+    //MARK: Create Variable
     let title: String
     let dateSubTitle: String
     let nameSubTitle: String
@@ -19,7 +19,7 @@ final class StandartBlogCellViewModel: StandartBlogCellIdentifiable {
         "StandartBlogTableViewCell"
     }
     
-    var cellHeight: CGFloat {
+    var cellHeight: Float {
         168
     }
     
@@ -29,10 +29,8 @@ final class StandartBlogCellViewModel: StandartBlogCellIdentifiable {
         dateSubTitle = blog.finalPost.date.stringDate()
         user = blog.user
     }
-    
 }
 
 final class StandartBlogSectionViewModel: StandartBlogSectionRowPresentable {
     var rows: [StandartBlogCellIdentifiable] = []
-    
 }
